+++ javascript/core/test/ibf.js
@@ -4,6 +4,8 @@
   var assert = require('assert');
   var ibf = require('../src/ibf');
   var utils = require('./utils');
+  var Readable = require('stream').Readable;
+  var util = require('util');
 
   var item1 = new Int8Array([5]).buffer;
   var item2 = new Int8Array([6]).buffer;
@@ -29,12 +31,27 @@
     }
   }
 
-  function* items1and2generator() {
+  function ArrayStream(items) {
+    Readable.call(this, { objectMode: true });
+    this._index = 0;
+    this._items = items;
+  }
+  util.inherits(ArrayStream, Readable);
+  ArrayStream.prototype._read = function() {
+    if (this._index >= this._items.length) {
+      this.push(null);
+    } else {
+      this.push(this._items[this._index]);
+      this._index++;
+    }
+  };
+
+  function* generator1and2() {
     yield item1;
     yield item2;
   }
 
-  function* items2and3generator() {
+  function* generator2and3() {
     yield item2;
     yield item3;
   }
@@ -43,15 +60,23 @@
   var just1 = empty.plus(item1);
   var just2 = empty.plus(item2);
   var just3 = empty.plus(item3);
-  var items1and2;
-  var items2and3;
+  var iterator1and2;
+  var iterator2and3;
+  var stream1and2;
+  var stream2and3;
 
   before(function(done) {
-    empty.plusIterator(items1and2generator()).then(function (i1a2) {
-      items1and2 = i1a2;
-      return empty.plusIterator(items2and3generator());
+    empty.plusIterator(generator1and2()).then(function (i1a2) {
+      iterator1and2 = i1a2;
+      return empty.plusIterator(generator2and3());
     }).then(function (i2a3) {
-      items2and3 = i2a3;
+      iterator2and3 = i2a3;
+      return empty.plusStream(new ArrayStream([item1, item2]));
+    }).then(function (s1a2) {
+      stream1and2 = s1a2;
+      return empty.plusStream(new ArrayStream([item2, item3]));
+    }).then(function (s2a3) {
+      stream2and3 = s2a3;
     }).then(done, done);
   });
 
@@ -90,11 +115,19 @@
         utils.assertThatSetOfArrayEquals(difference.added, [item3]);
         assert.deepEqual([], difference.removed);
 
-        difference = items1and2.toDifference();
+        difference = iterator1and2.toDifference();
         utils.assertThatSetOfArrayEquals(difference.added, [item1, item2]);
         assert.deepEqual([], difference.removed);
 
-        difference = items2and3.toDifference();
+        difference = iterator2and3.toDifference();
+        utils.assertThatSetOfArrayEquals(difference.added, [item2, item3]);
+        assert.deepEqual([], difference.removed);
+
+        difference = stream1and2.toDifference();
+        utils.assertThatSetOfArrayEquals(difference.added, [item1, item2]);
+        assert.deepEqual([], difference.removed);
+
+        difference = stream2and3.toDifference();
         utils.assertThatSetOfArrayEquals(difference.added, [item2, item3]);
         assert.deepEqual([], difference.removed);
       });
@@ -112,11 +145,19 @@
         utils.assertThatSetOfArrayEquals(difference.added, [item3]);
         assert.deepEqual([], difference.removed);
 
-        difference = goThroughJson(items1and2).toDifference();
+        difference = goThroughJson(iterator1and2).toDifference();
         utils.assertThatSetOfArrayEquals(difference.added, [item1, item2]);
         assert.deepEqual([], difference.removed);
 
-        difference = goThroughJson(items2and3).toDifference();
+        difference = goThroughJson(iterator2and3).toDifference();
+        utils.assertThatSetOfArrayEquals(difference.added, [item2, item3]);
+        assert.deepEqual([], difference.removed);
+
+        difference = goThroughJson(stream1and2).toDifference();
+        utils.assertThatSetOfArrayEquals(difference.added, [item1, item2]);
+        assert.deepEqual([], difference.removed);
+
+        difference = goThroughJson(stream2and3).toDifference();
         utils.assertThatSetOfArrayEquals(difference.added, [item2, item3]);
         assert.deepEqual([], difference.removed);
       });
@@ -137,11 +178,21 @@
         assert.deepEqual([], difference.added);
         utils.assertThatSetOfArrayEquals(difference.removed, [item3]);
 
-        empty.minusIterator(items1and2generator()).then(function (r) {
+        empty.minusIterator(generator1and2()).then(function (r) {
           difference = r.toDifference();
           assert.deepEqual([], difference.added);
           utils.assertThatSetOfArrayEquals(difference.removed, [item1, item2]);
-          return empty.minusIterator(items2and3generator());
+          return empty.minusIterator(generator2and3());
+        }).then(function (r) {
+          difference = r.toDifference();
+          assert.deepEqual([], difference.added);
+          utils.assertThatSetOfArrayEquals(difference.removed, [item2, item3]);
+          return empty.minusStream(new ArrayStream([item1, item2]));
+        }).then(function (r) {
+          difference = r.toDifference();
+          assert.deepEqual([], difference.added);
+          utils.assertThatSetOfArrayEquals(difference.removed, [item1, item2]);
+          return empty.minusStream(new ArrayStream([item2, item3]));
         }).then(function (r) {
           difference = r.toDifference();
           assert.deepEqual([], difference.added);
@@ -162,11 +213,21 @@
         assert.deepEqual([], difference.added);
         utils.assertThatSetOfArrayEquals(difference.removed, [item3]);
 
-        empty.minusIterator(items1and2generator()).then(function (r) {
+        empty.minusIterator(generator1and2()).then(function (r) {
           difference = goThroughJson(r).toDifference();
           assert.deepEqual([], difference.added);
           utils.assertThatSetOfArrayEquals(difference.removed, [item1, item2]);
-          return empty.minusIterator(items2and3generator());
+          return empty.minusIterator(generator2and3());
+        }).then(function (r) {
+          difference = goThroughJson(r).toDifference();
+          assert.deepEqual([], difference.added);
+          utils.assertThatSetOfArrayEquals(difference.removed, [item2, item3]);
+          return empty.minusStream(new ArrayStream([item1, item2]));
+        }).then(function (r) {
+          difference = goThroughJson(r).toDifference();
+          assert.deepEqual([], difference.added);
+          utils.assertThatSetOfArrayEquals(difference.removed, [item1, item2]);
+          return empty.minusStream(new ArrayStream([item2, item3]));
         }).then(function (r) {
           difference = goThroughJson(r).toDifference();
           assert.deepEqual([], difference.added);